mkjffs2: add lzo support
This adds support for building lzo images for root (and updates the
kernel accordingly, as LZO support wasn't compiled into the jffs2
driver).  However, I'm not sold on LZO, so I haven't enabled it by
default.  According to my tests, zlib left 540MB free, and bootup took
97s, while lzo left 505MB free, and bootup took 91s.  Program startup
time appeared to be about the same.  I'm not convinced that the 6s
bootup time improvement is worth it.

Signed-off-by: Andres Salomon <dilinger@debian.org>