whitespace in OFW updater script
While producing
http://wiki.laptop.org/go/User:Erik_Garrison/Image_Digestor I discovered
that the OFW updater script produced by the xodist tools had trailing
whitespace after the lines.

I rectified this by applying the attached patch to the line in question.
After application I verified that the output of the bios-crypto 'hashfs'
C script is byte-for-byte identical when run over the same input (in
this case the os767.img).

Signed-off-by: Andres Salomon <dilinger@debian.org>