installer: add actual guts..
This just does a simple dd to copy the entire drive, partition tables and all.
Obviously, this means if you're installing from a 1GB image onto a 60GB drive,
the drive will appear as only a 1GB drive until the tables/filesystem have
been resized.  This is something we can tweak later..

Signed-off-by: Andres Salomon <dilinger@collabora.co.uk>