debxo: drop hal keymapping stuff
Keymappings for OLPC are now handled by udev (and included upstream). In order
for this to work, OFW must have DMI support.  Newer versions of OFW all have it,
so upon detection of an OLPC machine, udev should automatically load the correct
keymap (and it should work properly for both X and console).

If using an older OFW that lacks DMI support, upgrading is recommended.  If
that's not an option, just add something like the following to your
/etc/rc.local (this will force manual loading of the keymaps):

/lib/udev/keymap /dev/input/event0 olpc-xo

Signed-off-by: Andres Salomon <dilinger@queued.net>