debxo: update UBIFS-related stuff for generic kernels
Stock Debian 3.2 kernels now run on XO-1 without modification (finally!).
However, they require some mount and boot option changes for our custom
hardware.

Since OFW does not support JFFS2 with LZO compression, we need to be
sure to mount /boot with compr=zlib so that any blocks written out to
the partition use zlib instead of lzo.

The Debian kernels also compile the redboot directory number to -1, so
we have to override that through a boot argument.

Note that a newer version of olpc-xo1-hw (0.3) is necessary to ensure
that all needed kernel modules are included in the initramfs images.

Signed-off-by: Andres Salomon <dilinger@queued.net>